+++ domain/executor.go
@@ -2,11 +2,16 @@ package domain
 
 import (
 	"bytes"
+	"errors"
 	"html/template"
 	"math/rand"
 	"regexp"
+	"strconv"
+	"time"
 
+	"github.com/google/uuid"
 	"github.com/valyala/fasthttp"
+	"github.com/wosai/deepmock/misc"
 )
 
 const (
@@ -20,6 +25,10 @@ const (
 	FilterModeRegular FilterMode = "regular"
 
 	ModeField = "mode"
+)
+
+var (
+	defaultTemplateFuncs template.FuncMap
 )
 
 type (
@@ -306,4 +315,73 @@ func (exe *Executor) FindRegulationExecutor(request *fasthttp.Request) *Regulati
 		}
 	}
 	return reg
+}
+
+// RegisterTemplateFunc 注册模板自定义函数
+func RegisterTemplateFunc(name string, f interface{}) error {
+	if _, ok := defaultTemplateFuncs[name]; ok {
+		return errors.New("func named " + name + " was exists")
+	}
+	defaultTemplateFuncs[name] = f
+	return nil
+}
+
+func genUUID() string {
+	return uuid.New().String()
+}
+
+func currentTimestamp(precision string) int64 {
+	now := time.Now().UnixNano()
+	switch precision {
+	case "mcs":
+		return now / 1e3
+	case "ms":
+		return now / 1e6
+	case "sec":
+		return now / 1e9
+	default:
+		return now
+	}
+}
+
+func formatDate(layout string) string {
+	return time.Now().Format(layout)
+}
+
+func plus(v interface{}, i int) interface{} {
+	switch v.(type) {
+	case int:
+		return v.(int) + i
+	case float64:
+		return v.(float64) + float64(i)
+	case float32:
+		return v.(float32) + float32(i)
+	case string:
+		vv, err := strconv.Atoi(v.(string))
+		if err != nil {
+			return "unsupported type"
+		}
+		return vv + i
+	default:
+		return "unsupported type"
+	}
+}
+
+func dateDelta(date, layout string, year, month, day int) string {
+	t, err := time.Parse(layout, date)
+	if err != nil {
+		return date
+	}
+	return t.AddDate(year, month, day).Format(layout)
+}
+
+func init() {
+	// create build-in template functions
+	defaultTemplateFuncs = make(template.FuncMap)
+	_ = RegisterTemplateFunc("uuid", genUUID)
+	_ = RegisterTemplateFunc("timestamp", currentTimestamp)
+	_ = RegisterTemplateFunc("date", formatDate)
+	_ = RegisterTemplateFunc("plus", plus)
+	_ = RegisterTemplateFunc("rand_string", misc.GenRandomString)
+	_ = RegisterTemplateFunc("date_delta", dateDelta)
 }
